refactor: depart interaction and renderer
